Refactor the data table
