+++ src/components/users/Table.tsx
@@ -1,6 +1,17 @@
-import { type Table as ReactTable, flexRender } from '@tanstack/react-table'
+import LoadingScreen from '~/components/common/LoadingScreen'
+import { trpc } from '~/utils/trpc'
+import {
+  type Table as ReactTable,
+  useReactTable,
+  getCoreRowModel,
+  getFilteredRowModel,
+  getPaginationRowModel,
+  type ColumnDef,
+  flexRender,
+} from '@tanstack/react-table'
+import { useMemo, useState } from 'react'
 
-export type User = {
+type User = {
   name: string | null
   roles: string | null
   department: string | null
@@ -8,7 +19,7 @@ export type User = {
   id: string
 }
 
-export const TableHeader = ({ table }: { table: ReactTable<User> }) => {
+const TableHeader = ({ table }: { table: ReactTable<User> }) => {
   return (
     <thead>
       {table.getHeaderGroups().map((headerGroup) => (
@@ -31,7 +42,7 @@ export const TableHeader = ({ table }: { table: ReactTable<User> }) => {
   )
 }
 
-export const TableBody = ({ table }: { table: ReactTable<User> }) => {
+const TableBody = ({ table }: { table: ReactTable<User> }) => {
   return (
     <tbody className="text-center">
       {table.getRowModel().rows.map((row) => {
@@ -49,4 +60,86 @@ export const TableBody = ({ table }: { table: ReactTable<User> }) => {
       })}
     </tbody>
   )
+}
+
+const Table = ({
+  data,
+  columns,
+}: {
+  data: User[]
+  columns: ColumnDef<User>[]
+}) => {
+  const table = useReactTable({
+    data,
+    columns,
+    getCoreRowModel: getCoreRowModel(),
+    getFilteredRowModel: getFilteredRowModel(),
+    getPaginationRowModel: getPaginationRowModel(),
+  })
+
+  return (
+    <table>
+      <TableHeader table={table} />
+      <TableBody table={table} />
+    </table>
+  )
+}
+
+/**
+ * @returns {JSX.Element} The data table consisting of the data
+ */
+export default function DataTable() {
+  const [users, setUsers] = useState<User[]>([])
+  const { isLoading } = trpc.user.getAllUsers.useQuery(undefined, {
+    onSuccess(data) {
+      setUsers(data)
+    },
+  })
+
+  const columns = useMemo<ColumnDef<User>[]>(
+    () => [
+      {
+        header: 'Users',
+        columns: [
+          {
+            accessorKey: 'id',
+            cell: (info) => info.getValue(),
+            header: 'No.',
+            footer: (props) => props.column.id,
+          },
+          {
+            accessorKey: 'name',
+            cell: (info) => info.getValue(),
+            header: 'Name',
+            footer: (props) => props.column.id,
+          },
+          {
+            accessorKey: 'email',
+            cell: (info) => info.getValue(),
+            header: 'Email',
+            footer: (props) => props.column.id,
+          },
+          {
+            accessorKey: 'department',
+            cell: (info) => info.getValue(),
+            header: 'Department',
+            footer: (props) => props.column.id,
+          },
+          {
+            accessorKey: 'roles',
+            cell: (info) => info.getValue(),
+            header: 'Role',
+            footer: (props) => props.column.id,
+          },
+        ],
+      },
+    ],
+    []
+  )
+
+  if (isLoading) {
+    return <LoadingScreen />
+  }
+
+  return <Table data={users} columns={columns} />
 }
